refactor: polish --merge/--since implementation
- Rename resolve-orig-head.ts → resolve-ref.ts (filename matches export)
- Make --merge imply --fix-local (only sensible post-merge behavior)
- Simplify --diff-filter=ACMR → AM (R/C are dead with diff-tree)
- Tighten test assertions to match full output format
- Fix misleading "simulate" comments in merge tests
- Update README to reflect --merge implying --fix-local
